Fetch HOME directory after other debhelpers
The postinst script of fruitnanny-api relies on sysuser-helper to
create the required fruitnanny system user and its home directory.
sysuser injects scripts via "#DEBHELPER#". This means before the
injection point, the system user potentially does not exist. Hence, the
HOME directory cannot be fetched at this point. Doing so leads to an
empty variable and chown is called without a directory leading to
installation errors.

Fix this issue by fetching the home directory after the debhelpers.