Use systemd to run upgrade process
Instead of running the upgrade as child process of the fruitnanny-api,
use a systemd one-shot service to execute the upgrade. This gives the
benefit, that the upgrade process is only triggered once but multiple
upgrade processes can wait on the systemd service to finish.
Furthermore, the upgrade process survices an API service shutdown and
restart.